pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "structured-dendrite"
version = "0.1.0"
description = "A compact experiment framework for dendrite+soma sequence models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "datasets>=2.19.0",
  "einops>=0.8.0",
  "hydra-core>=1.3.2",
  "lightning>=2.4.0",
  "omegaconf>=2.3.0",
  "rich>=13.7.1",
  "tensorboard>=2.19.0",
  # On the user's Windows + Python 3.11 machine, install the exact CUDA build we smoke-tested.
  "torch @ https://download.pytorch.org/whl/cu130/torch-2.10.0%2Bcu130-cp311-cp311-win_amd64.whl ; platform_system == 'Windows' and platform_machine == 'AMD64' and python_version == '3.11'",
  "torchvision @ https://download.pytorch.org/whl/cu130/torchvision-0.25.0%2Bcu130-cp311-cp311-win_amd64.whl ; platform_system == 'Windows' and platform_machine == 'AMD64' and python_version == '3.11'",
  # On a typical remote Linux x86_64 + Python 3.11 server, install the matching exact CUDA build too.
  "torch @ https://download.pytorch.org/whl/cu130/torch-2.10.0%2Bcu130-cp311-cp311-manylinux_2_28_x86_64.whl ; platform_system == 'Linux' and platform_machine == 'x86_64' and python_version == '3.11'",
  "torchvision @ https://download.pytorch.org/whl/cu130/torchvision-0.25.0%2Bcu130-cp311-cp311-manylinux_2_28_x86_64.whl ; platform_system == 'Linux' and platform_machine == 'x86_64' and python_version == '3.11'",
  # Fallback for other platforms, Python versions, or custom CUDA installs.
  "torch>=2.10,<2.11 ; (platform_system != 'Windows' or platform_machine != 'AMD64' or python_version != '3.11') and (platform_system != 'Linux' or platform_machine != 'x86_64' or python_version != '3.11')",
  "torchvision>=0.25,<0.26 ; (platform_system != 'Windows' or platform_machine != 'AMD64' or python_version != '3.11') and (platform_system != 'Linux' or platform_machine != 'x86_64' or python_version != '3.11')",
  "torchmetrics>=1.4.0",
  "transformers>=4.44.0",
]

[project.optional-dependencies]
research = [
  "flash-linear-attention>=0.2.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["structured_dendrite*"]
